refactor: travelEssential 패키지 분리
