+++ src/main/java/com/travel/role/domain/travelessential/dto/request/TravelEssentialDeleteReqDTO.java
@@ -1,4 +1,4 @@
-package com.travel.role.domain.room.dto;
+package com.travel.role.domain.travelessential.dto.request;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -10,8 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class TravelEssentialCheckReqDTO {
+public class TravelEssentialDeleteReqDTO {
 
-	private Boolean check;
 	private List<Long> ids = new ArrayList<>();
 }
