+++ src/main/java/com/travel/role/domain/travelessential/dto/request/TravelEssentialReqDTO.java
@@ -1,4 +1,4 @@
-package com.travel.role.domain.room.dto;
+package com.travel.role.domain.travelessential.dto.request;
 
 import static com.travel.role.global.exception.dto.ExceptionMessage.*;
 
@@ -8,9 +8,9 @@ import java.util.stream.Collectors;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 
-import com.travel.role.domain.room.entity.EssentialCategory;
+import com.travel.role.domain.travelessential.entity.EssentialCategory;
 import com.travel.role.domain.room.entity.Room;
-import com.travel.role.domain.room.entity.TravelEssentials;
+import com.travel.role.domain.travelessential.entity.TravelEssential;
 import com.travel.role.domain.user.entity.User;
 
 import lombok.AllArgsConstructor;
@@ -28,10 +28,10 @@ public class TravelEssentialReqDTO {
 	@Valid
 	private List<TravelEssentialItemDTO> items;
 
-	public List<TravelEssentials> toTravelEssentials(User user, Room room) {
+	public List<TravelEssential> toTravelEssentials(User user, Room room) {
 
 		return this.items.stream()
-			.map(item -> TravelEssentials.builder()
+			.map(item -> TravelEssential.builder()
 				.user(user)
 				.room(room)
 				.category(this.category)
