+++ src/main/java/com/travel/role/domain/travelessential/entity/TravelEssential.java
@@ -1,5 +1,20 @@
-package com.travel.role.domain.room.entity;
+package com.travel.role.domain.travelessential.entity;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+
+import org.hibernate.annotations.ColumnDefault;
+
+import com.travel.role.domain.room.entity.Room;
 import com.travel.role.domain.user.entity.User;
 
 import lombok.AllArgsConstructor;
@@ -7,20 +22,16 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
-
-import org.hibernate.annotations.ColumnDefault;
-
-@Table(name = "travel_essentials")
+@Table(name = "travel_essential")
 @Getter
 @Entity
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class TravelEssentials {
+public class TravelEssential {
 
 	@Id
-	@Column(name = "travel_essentials_id")
+	@Column(name = "travel_essential_id")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
